Fixed tab functionality to make Hash button on Hash Text tag become the default button when this tab is active.  The comparable buttons were already set to behave this way on the other tabs.
Added functionality to save the most recent settings to the Windows registry and restore them the next time the program is opened.  If the registry is not available (say we're running under Mono on Linux), it should gracefully return the defaults every time.

Added checkbox to the GUI to turn tooltips help on and off.  I think they're great for first time users (and that's why I put them in) but they get annoying later when you know what you're doing.