Added hooks to have SubWCRev automatically add revision and copyright info to various files (namely AssemblyInfo.cs and InnoSetup files).  The *.template files contain SubWCRev tags.  Make sure to add post-commit and post-update hooks to all working copies so SubWCRev can update the "real" version of these templates when changes occur.
Also updated the various program files so that displayed copyright information is now taken from the assembly info, which will in turn be updated by SubWCRev.  In other words, the displayed copyright year as well as the copyright year in the assembly metadata is now based on the last SVN revision and will be automatically updated.  Note that this will *not* update the year in source code comments; only the year that eventually gets displayed to the user in some fashion.