Restructured all three command-line utilities by creating a new utility class, CmdLineAppUtils, inside the WinHasherCore library. This class now contains the core code for all three programs, which now call the core to do their work. Since the only real difference between the programs was that md5 and sha1 hard-code the hash algorithm to MD5 and SHA-1 respectively, while hash lets the user pick the algorithm, it made sense to consolidate the code with flags and switches to accomodate the differences.
In the process, I took the chance to address Issue 2, "Output redirection". As discussed in that issue, dealing with Windows' standard command-line file redirection was becoming a serious pain as I was forced to parse the command line and handle the redirection myself. Instead, I added to the command-line core (which subsequently enables this functionality in all three programs) the "-in", "-out", "-append", and "-compare" command-line switches.

* "-in" allows the user to specify a input file which will contain a list of files to hash or compare. The files should be one per line. White space at the beginning and end of the line will be trimmed. Empty lines and lines containing only white space will be ignored. Lines beginning with a hash/pound symbol (#) will also be ignored and can be used for comments. If files are also specified on the command line, the input file takes precedence and the files on the command line are ignored.
* "-out" subsequently allows the output of the program to be directed to a file. By default, the output file is overwritten if it already exists.
* "-append" when used in conjunction with "-out" appends the output to an existing file rather than overwrites it. If the file does not exist, it will be created anyway, so this switch is safe to use even if the file does not exist.
* "-compare" puts the programs into comparison mode. See below for further details.

Based on these changes, I'm also changing the output format for the command-line programs. To bring WinHasher more in line with the UNIX utilities md5sum, shasum, etc., (see Issue 1), by default the command-line programs now output their results in a format similar to these UNIX programs, with the hash result first, followed by a delimiter, then the name of the file as passed in. Consequentially, if multiple files are specified, the programs no longer automatically perform a comparison; rather, they hash each file and output the result, much like the UNIX programs. Thus, the "-compare" flag has been added to restore the old functionality and compare multiple files, returning a simple Boolean match/no-match result. Obviously, if "-compare" is specified and only one file listed, this generates an error.

I'm probably going to add another flag, "-verify" that will read md5sum-style formats, hash the files in question, and generate a report. This will lay the groundwork for adding this functionality into the GUI app for Issue 1 and provide a stop-gap measure for those not afraid of the command line.

I've run out of time for coding and testing for today, so I'll update this progress as I move along.

Update issue 1
Update issue 2