SVN Clean-up and Bouncy Castle Conversion
This commit cleans up a few bits of cruft from our conversion from
Subversion to git, like removing the batch files and template files that
added SVN revision numbers to the application version numbers. git
doesn't have a concept like revision numbers; well, it does, but not
something that would be easy to add to a version number like the SVN
revisions. While I liked that convention, I'm afraid there's no easy
analog I can use from here on out.

I also started the transition from using Microsoft's Crypto API for the
hashing routines to the Legion of the Bouncy Castle. By moving to the
Bouncy Castle code, we can (a) add more hashing functions immediately
and (b) be better prepared to add new hashes (like SHA3) once the Bouncy
Castle folks release future updates. For now, the new hashes added are
SHA-224, RIPEMD-128, RIPEMD-256, RIPEMD-320, AND GOST3411. (While Bouncy
Castle supports MD2 and MD4, I elected not to support them in WinHasher;
I see MD5 being removed eventually as well.) This also means that the
internal implementations of Tiger and Whirlpool are no longer needed and
have been removed; both are already supported by the Bouncy Castle
library.

Also did a little bit of rearranging to consolidate a few things, like
converting hashes and output types to their display strings (and back)
within the WinHasherCore.HashEngine class rather than in a dozen
different places. This should also make it easier to add new hashes
later.

I also added changes to support using the main WinHasher app in
"portable" mode, meaning it will not make reads or writes to the
Registry while in GUI mode. To launch WinHasher in portable mode, add
the "-portable" command-line parameter on launch. Note that while the
SendTo shortcuts technically run in GUI mode, they don't launch the full
GUI and do not read or write to the Registry. None of the command-line
apps touch the Registry either, so the portable flag only affects the
main GUI app.

None of this really applies to any of the open issues, so there's still
a good bit left to tweak. The only issue touched here is #8; I removed
the private showMD5warning flag and all of its references, so the MD5
warning should never be displayed now. I'm not sure I'd consider this a
"fix", but I'll concede that it's an annoyance. We'll leave the warnings
to the documentation from here on out.