Setup and Registry Changes
Took some time to overhaul the InnoSetup script. Added the Bouncy Castle
library and the PathTweaker console app, both of which were accidentally
omitted. I also removed all the Send To shortcut icons, since I found
out there's no way to address Issue #7 within the installer. (I'll
probably need to update the app to give the user the option to add those
shortcuts.) I'm removing the md5 and sha1 executables and replacing them
with batch files that call the hash console app behind the scenes, then
added additional batch files for the other hashes as well.

In the main app, I'm restructuring the registry settings a bit and
switching the hash and output type settings for an index value to a
string. This will future-proof the drop-downs a bit when we add future
hashes, as we'll no longer be dependent on the order the hashes are in
the drop-down. This means changes to the UpgradeSettings() method, the
MainForm constructor, and the MainForm_FormClosing() method; basically
anywhere the settings are written to or read from the registry.