Add update checker and start Options dialog
Added GPFUpdateChecker to the project to add support for automatic
update checks.  This isn't complete, of course, but at least it's a
start.

Also started adding an Options dialog to allow the user to disable
automatic update checks and to restore the ability to add Send To
shortcuts, at least for the user running the app.  The Options dialog
will be disabled by default when WinHasher is run in portable mode.

Various other minor edits, such as tweaked comments, error messages,
validation checks, and an updated README.txt.  One significant addition
was tweaking the validation of the Compare To text boxes to trim leading
and trailing whitespace before comparing it to the generated hash (a pet
peeve of mine).