Fleshed out Options Dialog
Added the new Options Dialog to (a) give the user the ability to disable
automatic update checks and perform update checks on demand, and (b)
replace the Send To shortcuts that were created by the installer with an
option to let the user create them via the app itself. This second
option should let any user on the system create their own Send To
shortcuts, regardless of who runs the installer. (See Issue #7.)  The
options dialog is not available in portable mode.

Tweaked portable mode to actually display in the title bar that the app
is actually running in this mode.  Also fixed the default options to
correctly default the hash and output type drop-downs when there's no
existing preferences (or the user is running the app in portable mode).

Also added the GPF Update Checker XML schema definition file, which I
previously forgot.  The update checker won't run without it.  Oops...