Text Encoding Fix for Mono
Fixed the text encoding drop-down on the Hash Text tab to work around a
known Mono bug. Mono has a bug in Encoding.GetEncodings() that returns a
bunch of encodings that are invalid for the current platform, and when
you try to add those to the combo box, this throws an exception. This
doesn't happen under Microsoft's .NET implementation. I fixed this in
Cryptnos when I started targeting Mono as a platform, so I might as well
do it for WinHasher as well. Note that this also changes the
TextEncoding registry key from a DWORD (an index in the combo box list)
to a String so we can more reliably look it up if the list somehow
changes. I also switched the default encoding in the drop-down to UTF-8
from the system default, mostly because it should be a safer choice for
multiplatform deployments.