BC 1.8.1 Upgrade, Decommission md5 and sha1 console apps
I'm upgrading the Bouncy Castle library to 1.8.1, which is the latest
version. The primary advantage here is the addition of the SHA3 hashes.
The 224, 256, 384, and 512 bit length versions have already been added.
Tentative support for SHAKE is also planned but currently commented out;
I'm not sure I get how the BC implementation is set up, because it
doesn't quite match Wikipedia's explanation of how it works. I'll need
to research that a bit before I decide to add it.

I'm also going ahead and decommissioning the old md5 and sha1 console
apps, replacing them with the convenience batch files previously
created. That's two less code branches to maintain.

On the main app's Hash Text form, I also lifted the restriction that
prevented it from generating a hash if the Input Text field was empty.
Technically, an empty input value is a valid test vector for any hash,
so it seems silly not to allow that.

I'm contemplating making this a 2.0 release rather than a 1.7 release,
given the number of changes involved. I still need to update the the
documentation for all the above changes.